requirements.txt
django-phonenumber-field[phonenumberslite]
Pillow
stripe
django-model-utils
cryptography
